refactor(database): 将数据库初始化逻辑统一整合到 database 包 - 新增 app/database/__init__.py，整合建表、种子数据和基线计算为 init_database() - 将 scripts/seed_data.py 和 scripts/compute_baseline.py 迁移到 app/database/ - 删除 scripts/init_db.py，建表逻辑已合并到 database 包 - 种子数据和基线计算增加幂等性检查，仅在空表时执行 - main.py 启动时调用 init_database() 替代内联建表代码 - 使用 pathlib.Path 替代 os.path 处理路径 - 简化 CI 流程和部署脚本中的数据库初始化步骤
